Correction for GCC compatibility.
git-svn-id: http://b035a0a.online-server.cloud/svn/Acf@626 5144ca80-2444-0410-90ef-8152f95a0219

+++ Include/iqt/CDateTime.cpp
@@ -134,7 +134,7 @@ double CDateTime::ToCTime() const
 void CDateTime::FromCTime(double ctime)
 {
 	unsigned ctimeValue = unsigned(ctime);
-	setTime_t(ctime);
+	setTime_t(uint(ctime));
 	QDateTime::operator=(addMSecs(qint64((ctime - ctimeValue) * 1000)));
 }
